requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.11.0
filelock==3.13.1
Flask==3.1.1
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.6.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
idna==3.10
ipykernel==6.29.5
ipython==8.36.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.65
langchain-ollama==0.3.3
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.45
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.2.5
ollama==0.5.1
openai==1.78.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
pip==25.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.4
pydantic_core==2.33.2
pydantic-settings==2.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
torchvision==0.22.0+cu128
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.0
zstandard==0.23.0
